refactor(settings_ui): extract tool metadata + visibility to _tools_meta.py
ToolStub/MANDATORY_TOOLS/FEATURE_GATED_TOOLS/_get_tool_metadata/
apply_tool_visibility and the capability-field helpers. No test
patch-through on these names, so callers use the re-exports unchanged.
